% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modCount.R
\name{modCount}
\alias{modCount}
\title{modules counter}
\usage{
modCount(url)
}
\arguments{
\item{url}{A nexusmods game-page url}
}
\value{
A number specifying the number of mods that exist for the target game
}
\description{
Function to obtain the number of mods that exist on nexusmods for a particular game
}
\examples{
modCount("https://www.nexusmods.com/devilmaycry5/mods/")
}
